move more functions to coords.c/.h
